refactor: refactor core components to the core
