+++ src/app/core/home/home-routing.module.ts
